changed

+++ test/Button.pde
@@ -1,12 +1,12 @@
 class Button{
   //width, height, start x, start y, button label/name
-  int w;
-  int h;
-  int x;
-  int y;
+  float w;
+  float h;
+  float x;
+  float y;
   
   //constructor
-  Button(int tempX, int tempY, int tempW, int tempH){
+  Button(float tempX, float tempY, float tempW, float tempH){
     x = tempX;
     y = tempY;
     w = tempW;
